refactor(7A.2): split partition memberships into their own repo
Restores the 7A.2 one-repo-per-entity layout: partition memberships
get their own port and asyncpg implementation instead of riding on
UserRepository.

- new core/ports/partition_membership_repo.py (PartitionMembershipRepository ABC)
- new services/persistence/partition_membership_repo.py (PgPartitionMembershipRepository)
- moved the 6 clean methods + 6 legacy shim methods + _row_to_user_partition
  off PgUserRepository
- UserRepository keeps reading partition_memberships via _fetch_memberships
  only to hydrate the User aggregate's partitions field (read-only
  denormalisation, not membership management)
- expose membership_repo on CatalogStore, PostgresStore and the DI container
- repoint the 7 membership call sites in the vectordb.py shim
- split the membership tests into tests/integration/test_partition_membership_repo.py

Unit suite unchanged (714 passed); no caller outside the shim touched.
