+++ openrag/core/ports/catalog_store.py
@@ -17,6 +17,7 @@ from .idempotency_repo import IdempotencyRepository
 from .job_repo import JobRepository
 from .model_endpoint_repo import ModelEndpointRepository
 from .oidc_session_repo import OIDCSessionRepository
+from .partition_membership_repo import PartitionMembershipRepository
 from .partition_repo import PartitionRepository
 from .preset_repo import PresetRepository
 from .prompt_repo import PromptRepository
@@ -56,6 +57,10 @@ class CatalogStore(ABC):
     @property
     @abstractmethod
     def partition_repo(self) -> PartitionRepository: ...
+
+    @property
+    @abstractmethod
+    def membership_repo(self) -> PartitionMembershipRepository: ...
 
     @property
     @abstractmethod
